move recs sidepanel into table component
