+++ packages/client/src/pages/RecommendationsPage/RecommendationsPage.test.tsx
@@ -3,7 +3,7 @@
  */
 
 import React from 'react'
-import { fireEvent, render, screen } from '@testing-library/react'
+import { render } from '@testing-library/react'
 import {
   EstimationResult,
   RecommendationResult,
@@ -16,7 +16,6 @@ import {
 } from '../../utils/hooks'
 import { ServiceResult } from '../../Types'
 import moment from 'moment'
-import { act } from 'react-dom/test-utils'
 
 jest.mock('../../utils/hooks/RecommendationsServiceHook')
 jest.mock('../../utils/hooks/FootprintServiceHook')
@@ -102,79 +101,6 @@ describe('Recommendations Page', () => {
     const { getByRole } = render(<RecommendationsPage />)
 
     expect(getByRole('progressbar')).toBeInTheDocument()
-  })
-
-  it('displays a selected recommendation in a side panel when its row is clicked', () => {
-    const { getByText, queryByTestId } = render(<RecommendationsPage />)
-
-    expect(queryByTestId('sideBarTitle')).toBeFalsy()
-
-    fireEvent.click(screen.getByText('test-a'))
-    const recommendationDetail = getByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(recommendationDetail).toBeInTheDocument()
-  })
-
-  it('hides and re-displays the side panel when a recommendation row is clicked multiple times', () => {
-    const { getByText, queryByTestId } = render(<RecommendationsPage />)
-
-    fireEvent.click(screen.getByText('test-a'))
-    fireEvent.click(screen.getByText('test-a', { selector: 'div' }))
-    // selector added because now there are multiple elements with the text 'test-a' on the screen
-
-    expect(queryByTestId('sideBarTitle')).toBeFalsy()
-
-    fireEvent.click(screen.getByText('test-a'))
-    const recommendationDetail = getByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(recommendationDetail).toBeInTheDocument()
-  })
-
-  it('displays a new recommendation in the side panel when its recommendation row is clicked', () => {
-    const { getByText, queryByText, queryByTestId } = render(
-      <RecommendationsPage />,
-    )
-
-    fireEvent.click(screen.getByText('test-a'))
-    fireEvent.click(screen.getByText('test-b'))
-    const firstRecommendationDetail = queryByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-    const secondRecommendationDetail = getByText(
-      mockRecommendationData[1].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(firstRecommendationDetail).toBeFalsy()
-    expect(secondRecommendationDetail).toBeInTheDocument()
-  })
-
-  it('re-displays the side panel when the close button and then the same recommendation row is clicked', () => {
-    const { getByText, getByTestId, queryByTestId } = render(
-      <RecommendationsPage />,
-    )
-
-    fireEvent.click(screen.getByText('test-a'))
-
-    const closeIcon = getByTestId('closeIcon')
-
-    act(() => {
-      fireEvent.click(closeIcon)
-    })
-
-    fireEvent.click(screen.getByText('test-a'))
-    const recommendationDetail = getByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(recommendationDetail).toBeInTheDocument()
   })
 
   it('should render a filter bar on the page', () => {
