+++ packages/client/src/pages/RecommendationsPage/RecommendationsTable/RecommendationsTable.test.tsx
@@ -14,6 +14,8 @@ import {
 } from '../../../utils/data'
 import RecommendationsTable from './RecommendationsTable'
 import { useRemoteFootprintService } from '../../../utils/hooks'
+import { act } from 'react-dom/test-utils'
+import React from 'react'
 
 jest.mock('../../../utils/hooks/FootprintServiceHook')
 
@@ -92,17 +94,93 @@ describe('Recommendations Table', () => {
     expect(actualRowData[1].length).toBe(expectedRowData[1].length - 5)
   })
 
-  it('calls handle row click with the recommendations details when a row is clicked', () => {
-    const handleRowClick = jest.fn()
-    render(
+  it('displays a selected recommendation in a side panel when its row is clicked', () => {
+    const { getByText, queryByTestId } = render(
       <RecommendationsTable
         {...testProps}
         recommendations={mockRecommendationData}
-        handleRowClick={handleRowClick}
       />,
     )
+
+    expect(queryByTestId('sideBarTitle')).toBeFalsy()
+
     fireEvent.click(screen.getByText('test-a'))
-    expect(handleRowClick).toHaveBeenCalledTimes(1)
+    const recommendationDetail = getByText(
+      mockRecommendationData[0].recommendationDetail,
+    )
+
+    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
+    expect(recommendationDetail).toBeInTheDocument()
+  })
+
+  it('hides and re-displays the side panel when a recommendation row is clicked multiple times', () => {
+    const { getByText, queryByTestId } = render(
+      <RecommendationsTable
+        {...testProps}
+        recommendations={mockRecommendationData}
+      />,
+    )
+
+    fireEvent.click(screen.getByText('test-a'))
+    fireEvent.click(screen.getByText('test-a', { selector: 'div' }))
+    // selector added because now there are multiple elements with the text 'test-a' on the screen
+
+    expect(queryByTestId('sideBarTitle')).toBeFalsy()
+
+    fireEvent.click(screen.getByText('test-a'))
+    const recommendationDetail = getByText(
+      mockRecommendationData[0].recommendationDetail,
+    )
+
+    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
+    expect(recommendationDetail).toBeInTheDocument()
+  })
+
+  it('displays a new recommendation in the side panel when its recommendation row is clicked', () => {
+    const { getByText, queryByText, queryByTestId } = render(
+      <RecommendationsTable
+        {...testProps}
+        recommendations={mockRecommendationData}
+      />,
+    )
+
+    fireEvent.click(screen.getByText('test-a'))
+    fireEvent.click(screen.getByText('test-b'))
+    const firstRecommendationDetail = queryByText(
+      mockRecommendationData[0].recommendationDetail,
+    )
+    const secondRecommendationDetail = getByText(
+      mockRecommendationData[1].recommendationDetail,
+    )
+
+    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
+    expect(firstRecommendationDetail).toBeFalsy()
+    expect(secondRecommendationDetail).toBeInTheDocument()
+  })
+
+  it('re-displays the side panel when the close button and then the same recommendation row is clicked', () => {
+    const { getByText, getByTestId, queryByTestId } = render(
+      <RecommendationsTable
+        {...testProps}
+        recommendations={mockRecommendationData}
+      />,
+    )
+
+    fireEvent.click(screen.getByText('test-a'))
+
+    const closeIcon = getByTestId('closeIcon')
+
+    act(() => {
+      fireEvent.click(closeIcon)
+    })
+
+    fireEvent.click(screen.getByText('test-a'))
+    const recommendationDetail = getByText(
+      mockRecommendationData[0].recommendationDetail,
+    )
+
+    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
+    expect(recommendationDetail).toBeInTheDocument()
   })
 
   it('shows undefined cell values as "-" within the table', () => {
