+++ packages/client/src/utils/hooks/index.ts
@@ -5,3 +5,5 @@
 export { default as useRemoteEmissionService } from './EmissionFactorServiceHook'
 export { default as useRemoteFootprintService } from './FootprintServiceHook'
 export { default as useRemoteRecommendationsService } from './RecommendationsServiceHook'
+export * from './FootprintDataHook'
+export * from './RecommendationsDataHook'
